requirements.txt
# Django==4.1.3
Django==3.1.7

djangorestframework==3.12.4

pytz==2021.3

django-cors-headers==3.7.0
django-environ==0.4.5

psycopg2-binary==2.9.5

Pillow==9.4.0

django-ckeditor==6.1.0

waitress

whitenoise